+++ config/migrations/01_create_log_table.sql
@@ -2,7 +2,7 @@ create schema if not exists payments;
 
 create table if not exists payments.log (
     id uuid primary key,
-    amount double precision not null,
+    amount decimal not null,
     requested_at timestamptz not null,
     processed_by text
 );
